nodos/requirements.txt
Pyro4>=4.82
Pillow>=9.0.0